refactor : final 사용
