+++ src/main/java/wooteco/subway/controller/PageController.java
@@ -10,8 +10,8 @@ import wooteco.subway.service.admin.LineService;
 
 @Controller
 public class PageController {
-	private LineService lineService;
-	private StationRepository stationRepository;
+	private final LineService lineService;
+	private final StationRepository stationRepository;
 
 	public PageController(LineService lineService, StationRepository stationRepository) {
 		this.lineService = lineService;
