+++ src/main/java/wooteco/subway/dto/PathResponse.java
@@ -3,34 +3,32 @@ package wooteco.subway.dto;
 import java.util.List;
 
 public class PathResponse {
-    private List<StationResponse> stations;
-    private int distance;
-    private int duration;
+	private List<StationResponse> stations;
+	private int distance;
+	private int duration;
 
-    public PathResponse() {
-    }
+	public PathResponse() {
+	}
 
-    public PathResponse(List<StationResponse> stations, int distance, int duration) {
-        this.stations = stations;
-        this.distance = distance;
-        this.duration = duration;
-    }
+	public PathResponse(List<StationResponse> stations, int distance, int duration) {
+		this.stations = stations;
+		this.distance = distance;
+		this.duration = duration;
+	}
 
-    public PathResponse(List<StationResponse> stations, double distance, double duration) {
-        this.stations = stations;
-        this.distance = (int)distance;
-        this.duration = (int)duration;
-    }
+	public PathResponse(List<StationResponse> stations, double distance, double duration) {
+		this(stations, (int)distance, (int)duration);
+	}
 
-    public List<StationResponse> getStations() {
-        return stations;
-    }
+	public List<StationResponse> getStations() {
+		return stations;
+	}
 
-    public int getDistance() {
-        return distance;
-    }
+	public int getDistance() {
+		return distance;
+	}
 
-    public int getDuration() {
-        return duration;
-    }
+	public int getDuration() {
+		return duration;
+	}
 }
