+++ src/main/java/wooteco/subway/repository/StationRepository.java
@@ -8,7 +8,7 @@ import wooteco.subway.domain.Station;
 
 public interface StationRepository extends CrudRepository<Station, Long> {
     @Override
-    List<Station> findAllById(Iterable ids);
+    List<Station> findAllById(Iterable<Long> ids);
 
     @Override
     List<Station> findAll();
