+++ src/main/java/wooteco/subway/service/admin/LineService.java
@@ -15,8 +15,8 @@ import wooteco.subway.repository.StationRepository;
 
 @Service
 public class LineService {
-    private LineRepository lineRepository;
-    private StationRepository stationRepository;
+    private final LineRepository lineRepository;
+    private final StationRepository stationRepository;
 
     public LineService(LineRepository lineRepository, StationRepository stationRepository) {
         this.lineRepository = lineRepository;
